backend/requirements.txt
flask
flask-cors
azure-storage-blob
azure-cosmos
